MAINT: Move _collapse_runs/_shorten_paths to _logging, comment the lock
